Split `data` module into `icebreaker_core` subcrate
