+++ core/src/model.rs
